+++ src/Security/src/Matcher.php
@@ -1,12 +1,5 @@
 <?php
 
-/**
- * Spiral Framework.
- *
- * @license   MIT
- * @author    Anton Titov (Wolfy-J)
- */
-
 declare(strict_types=1);
 
 namespace Spiral\Security;
@@ -22,28 +15,25 @@ final class Matcher
 {
     public function isPattern(string $string): bool
     {
-        return strpos($string, '*') !== false || strpos($string, '|') !== false;
+        return \str_contains($string, '*') || \str_contains($string, '|');
     }
 
     /**
      * Checks if string matches given pattent.
-     *
-     *
      */
     public function matches(string $string, string $pattern): bool
     {
-        if ($string === $pattern) {
-            return true;
-        }
-        if (!$this->isPattern($pattern)) {
-            return false;
-        }
-        return (bool)preg_match($this->getRegex($pattern), $string);
+        return match (true) {
+            $string === $pattern => true,
+            !$this->isPattern($pattern) => false,
+            default => (bool) \preg_match($this->getRegex($pattern), $string)
+        };
     }
 
     private function getRegex(string $pattern): string
     {
-        $regex = str_replace('*', '[a-z0-9_\-]+', addcslashes($pattern, '.-'));
-        return "#^{$regex}$#i";
+        $regex = \str_replace('*', '[a-z0-9_\-]+', \addcslashes($pattern, '.-'));
+
+        return \sprintf('#^%s$#i', $regex);
     }
 }