+++ src/Security/src/Rule/CallableRule.php
@@ -1,12 +1,5 @@
 <?php
 
-/**
- * Spiral Framework.
- *
- * @license   MIT
- * @author    Anton Titov (Wolfy-J)
- */
-
 declare(strict_types=1);
 
 namespace Spiral\Security\Rule;
@@ -19,21 +12,14 @@ use Spiral\Security\RuleInterface;
  */
 final class CallableRule implements RuleInterface
 {
-    /**
-     * @var callable
-     */
-    private $callable;
+    private string|array|object $callable;
 
-    public function __construct(callable $callable)
-    {
+    public function __construct(callable $callable) {
         $this->callable = $callable;
     }
 
-    /**
-     * {@inheritdoc}
-     */
     public function allows(ActorInterface $actor, string $permission, array $context): bool
     {
-        return (bool)call_user_func($this->callable, $actor, $permission, $context);
+        return (bool) \call_user_func($this->callable, $actor, $permission, $context);
     }
 }